# Add API to query polygons in batches