This adds a version of queryPolygon that takes an interface which is
passed batches of polygons in the search area. This new API is then used
from the old queryPolygons to collect polygons, and from findNearestRef
to find the nearest poly in an arbitrarily sized area. Previously
findNearestPoly had an arbitrary poly limit of 128 which could cause it
to return potentially wrong results.

Fix #107

I was concerned about the impact of this since it adds some virtual overhead
and additional copying to `queryPolygons`. I've added some benchmarks below.
At least on my CPU the overhead is not extremely high, though it is still around 10%
at large extents. The comments at the top of `DetourNavMeshQuery.h` say that
virtual dispatch is expensive on certain platforms, so the performance regression
may be worse on those platforms. We could have two versions of `queryPolygonsInTile`,
but of course I would prefer to keep it as it is now. Thoughts?

### Benchmarks for `findNearestPoly`
I varied extents up to 14, 28, 14. Tests at higher extents weren't meaningful due to
the old version reaching the 128 polygon limit and thus potentially returning wrong results.
#### New
Extents | Number of queries | Total time | Average time per query
--- | --- | --- | ---
(2, 4, 2) | 2710 | 1.49 ms | 0.550 microseconds
(4, 8, 4) | 2710 | 3.38 ms | 1.249 microseconds
(6, 12, 6) | 2710 | 6.08 ms | 2.245 microseconds
(8, 16, 8) | 2710 | 9.48 ms | 3.498 microseconds
(10, 20, 10) | 2710 | 13.48 ms | 4.973 microseconds
(12, 24, 12) | 2710 | 16.91 ms | 6.239 microseconds
(14, 28, 14) | 2710 | 20.64 ms | 7.615 microseconds

#### Old
Extents | Number of queries | Total time | Average time per query
--- | --- | --- | ---
(2, 4, 2) | 2710 | 1.46 ms | 0.538 microseconds
(4, 8, 4) | 2710 | 3.28 ms | 1.209 microseconds
(6, 12, 6) | 2710 | 6.06 ms | 2.235 microseconds
(8, 16, 8) | 2710 | 9.20 ms | 3.396 microseconds
(10, 20, 10) | 2710 | 12.97 ms | 4.785 microseconds
(12, 24, 12) | 2710 | 16.50 ms | 6.089 microseconds
(14, 28, 14) | 2710 | 20.14 ms | 7.430 microseconds

Almost no overhead in the new version.

### Benchmarks for `queryPolygons`
#### New
Extents | Number of queries | Total time | Average time per query
--- | --- | --- | ---
(2, 4, 2) | 2710  | 0.45 ms | 0.165 microseconds
(4, 8, 4) | 2710  | 0.79 ms | 0.292 microseconds
(6, 12, 6) | 2710  | 1.15 ms | 0.425 microseconds
(8, 16, 8) | 2710  | 1.60 ms | 0.590 microseconds
(10, 20, 10) | 2710 | 2.07 ms | 0.762 microseconds
(12, 24, 12) | 2710 | 2.33 ms | 0.861 microseconds
(14, 28, 14) | 2710 | 2.67 ms | 0.986 microseconds
(16, 32, 16) | 2710 | 3.15 ms | 1.163 microseconds
(18, 36, 18) | 2710 | 3.61 ms | 1.332 microseconds
(20, 40, 20) | 2710 | 4.04 ms | 1.492 microseconds
(22, 44, 22) | 2710 | 6.36 ms | 2.347 microseconds
(24, 48, 24) | 2710 | 5.09 ms | 1.877 microseconds
(26, 52, 26) | 2710 | 5.49 ms | 2.026 microseconds
(28, 56, 28) | 2710 | 5.92 ms | 2.183 microseconds
(30, 60, 30) | 2710 | 6.44 ms | 2.376 microseconds
(32, 64, 32) | 2710 | 6.99 ms | 2.578 microseconds
(34, 68, 34) | 2710 | 7.22 ms | 2.666 microseconds
(36, 72, 36) | 2710 | 7.70 ms | 2.842 microseconds
(38, 76, 38) | 2710 | 8.19 ms | 3.022 microseconds
(40, 80, 40) | 2710 | 8.59 ms | 3.170 microseconds
(42, 84, 42) | 2710 | 9.14 ms | 3.372 microseconds
(44, 88, 44) | 2710 | 9.44 ms | 3.483 microseconds
(46, 92, 46) | 2710 | 9.90 ms | 3.652 microseconds
(48, 96, 48) | 2710 | 10.08 ms | 3.718 microseconds
(50, 100, 50) | 2710 | 10.72 ms | 3.955 microseconds

### Old
Extents | Number of queries | Total time | Average time per query
--- | --- | --- | ---
(2, 4, 2) | 2710 | 0.38 ms | 0.141 microseconds
(4, 8, 4) | 2710 |  0.64 ms | 0.236 microseconds
(6, 12, 6) | 2710 | 0.97 ms | 0.358 microseconds
(8, 16, 8) | 2710 | 1.34 ms | 0.496 microseconds
(10, 20, 10) | 2710 | 1.71 ms | 0.632 microseconds
(12, 24, 12) | 2710 | 1.95 ms | 0.721 microseconds
(14, 28, 14) | 2710 | 2.40 ms | 0.886 microseconds
(16, 32, 16) | 2710 | 2.65 ms | 0.977 microseconds
(18, 36, 18) | 2710 | 3.04 ms | 1.120 microseconds
(20, 40, 20) | 2710 | 3.51 ms | 1.296 microseconds
(22, 44, 22) | 2710 | 3.87 ms | 1.427 microseconds
(24, 48, 24) | 2710 | 4.39 ms | 1.620 microseconds
(26, 52, 26) | 2710 | 4.65 ms | 1.717 microseconds
(28, 56, 28) | 2710 | 5.05 ms | 1.865 microseconds
(30, 60, 30) | 2710 | 5.59 ms | 2.063 microseconds
(32, 64, 32) | 2710 | 5.92 ms | 2.185 microseconds
(34, 68, 34) | 2710 | 6.25 ms | 2.305 microseconds
(36, 72, 36) | 2710 | 6.73 ms | 2.485 microseconds
(38, 76, 38) | 2710 | 7.13 ms | 2.631 microseconds
(40, 80, 40) | 2710 | 7.54 ms | 2.783 microseconds
(42, 84, 42) | 2710 | 7.86 ms | 2.900 microseconds
(44, 88, 44) | 2710 | 8.32 ms | 3.068 microseconds
(46, 92, 46) | 2710 | 8.68 ms | 3.201 microseconds
(48, 96, 48) | 2710 | 8.88 ms | 3.276 microseconds
(50, 100, 50) | 2710 | 9.47 ms | 3.495 microseconds