@hymerman 2016-01-26
This looks good! Thnaks! I'd accept that small extra overhead for the functionality that's gained - I remember this being a huge pain point in many ways!

I'll wait for more comments before merging.